+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Adapted to the Calliope 0.7.0 release
+- Python 3.14 is now supported
+
 ## 0.1.0 (2026-08-27)
 
 Initial release of Calliope Studio, a browser-based interface for [Calliope](https://calliope.readthedocs.io/) (version 0.7 and later).